+++ apps/client/src/app/components/user-account-access/create-or-update-access-dialog/interfaces/interfaces.ts
@@ -3,3 +3,6 @@ import { Access } from '@ghostfolio/common/interfaces';
 export interface CreateOrUpdateAccessDialogParams {
   access?: Access;
 }
+
+export type AccessLevel =
+  'CREATE_READ_UPDATE_DELETE' | 'READ' | 'READ_RESTRICTED';